Add files via upload
This a a method of display that you might find useful. Since we don't know how to pass information via cookies reliably yet, we can run the framework in JS based on the map given from the server. This file shows some of the logic required for the game loop. The display is poorly spaced and will need improvement (characters to be replaced with images, properly spaced, limited view where the icons around the character move instead of the cursor, properly aligned etc). Note that in the final version for multiplayer to work we will need to pass updated positions to the server. Field of view is yet to be determined, potentially 2-3 "spaces" in each direction.